+++ tests/e2e/tests/tags.js
